% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachCellID.R
\name{attachGridVars}
\alias{attachGridVars}
\title{Attach grid variables to dataset}
\usage{
attachGridVars(df, make_grid, lon_colname, lat_colname)
}
\arguments{
\item{df}{data.frame}

\item{make_grid}{list}

\item{lon_colname}{character}

\item{lat_colname}{character}
}
\value{
data.frame
}
\description{
Attach grid variables to dataset
}
