% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachCellID.R
\name{attachHexVars}
\alias{attachHexVars}
\title{Attach hexbins}
\usage{
attachHexVars(df, make_hex, lon_colname, lat_colname)
}
\arguments{
\item{df}{data.frame}

\item{make_hex}{SpatialPolygons}

\item{lon_colname}{character}

\item{lat_colname}{character}
}
\value{
data.frame
}
\description{
Attach hexbins
}
