% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachCellID.R
\name{attachShpVars}
\alias{attachShpVars}
\title{Attach custom shapefile cell ids to data}
\usage{
attachShpVars(df, shp_poly, poly_id_colname, lon_colname, lat_colname)
}
\arguments{
\item{df}{data.frame}

\item{shp_poly}{SpatialPolygons}

\item{poly_id_colname}{character}

\item{lon_colname}{character}

\item{lat_colname}{character}
}
\value{
data.frame
}
\description{
Attach custom shapefile cell ids to data
}
