% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bootstrap}
\alias{bootstrap}
\title{Produce a vector of n classes partitioned by a vector of breaks.
Also accepts part = 'equal' for n equal sized classes
and part = 'ml' that gives 60\%, 20\%, 20\% partitions intended for test, training, and validation sets.}
\usage{
bootstrap(df, part, n = 3)
}
\arguments{
\item{df}{data.frame}

\item{part}{numeric}

\item{n}{integer}
}
\value{
integer
}
\description{
Produce a vector of n classes partitioned by a vector of breaks.
Also accepts part = 'equal' for n equal sized classes
and part = 'ml' that gives 60\%, 20\%, 20\% partitions intended for test, training, and validation sets.
}
