% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cantorPair}
\alias{cantorPair}
\title{Combine x and y location columns into a single identifier column for each grid cell.}
\usage{
cantorPair(xy)
}
\arguments{
\item{xy}{matrix}
}
\value{
numeric
}
\description{
Combine x and y location columns into a single identifier column for each grid cell.
}
