% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cellCoords}
\alias{cellCoords}
\title{Convert grid cell x/y identifiers and return the original coordinate values of the centre of each grid cell}
\usage{
cellCoords(xy, origin, cellsize)
}
\arguments{
\item{xy}{matrix}

\item{origin}{numeric}

\item{cellsize}{numeric}
}
\value{
matrix
}
\description{
Convert grid cell x/y identifiers and return the original coordinate values of the centre of each grid cell
}
