% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cellXY}
\alias{cellXY}
\title{Convert lon/lat values into grid cell row/col (zero indexed, integer) identifiers using an origin point and cell width/height}
\usage{
cellXY(dat, origin, cellsize)
}
\arguments{
\item{dat}{data.frame}

\item{origin}{numeric}

\item{cellsize}{numeric}
}
\value{
matrix
}
\description{
Convert lon/lat values into grid cell row/col (zero indexed, integer) identifiers using an origin point and cell width/height
}
