% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstGenModel}
\alias{firstGenModel}
\title{Run the first generation of models}
\usage{
firstGenModel(df, cell_ids, model, n_class)
}
\arguments{
\item{df}{data.frame}

\item{cell_ids}{integer}

\item{model}{formula}

\item{n_class}{integer}
}
\value{
list
}
\description{
Run the first generation of models
}
