% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachCellID.R
\name{makeGrid}
\alias{makeGrid}
\title{Make grid object}
\usage{
makeGrid(df, lon_colname, lat_colname, resolution, square = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{lon_colname}{character}

\item{lat_colname}{character}

\item{resolution}{integer}

\item{square}{boolean}
}
\value{
list
}
\description{
Make grid object
}
