% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOrganismModels.R
\name{makeOrganismModels}
\alias{makeOrganismModels}
\title{Return models and bounds for input into runGeneAlgorithm}
\usage{
makeOrganismModels(gdf, response, control_vars = c())
}
\arguments{
\item{gdf}{data.frame}

\item{response}{character}

\item{control_vars}{character}
}
\value{
list
}
\description{
Return models and bounds for input into runGeneAlgorithm
}
