% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot}

\item{plotlist}{list}

\item{cols}{integer}

\item{layout}{matrix}
}
\description{
Multiple plot function
}
