% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutateGen}
\alias{mutateGen}
\title{Mutate the offspring}
\usage{
mutateGen(
  genome,
  n_class,
  mute_intensity,
  n_gen = NA,
  gen = NA,
  mute_decay = 2
)
}
\arguments{
\item{genome}{integer}

\item{n_class}{integer}

\item{mute_intensity}{numeric}

\item{n_gen}{numeric}

\item{gen}{numeric}

\item{mute_decay}{numeric}
}
\value{
integer
}
\description{
Mutate the offspring
}
