% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nGenModel}
\alias{nGenModel}
\title{Run the nth generation of models}
\usage{
nGenModel(df, cell_ids, model, n_class, genome, gen)
}
\arguments{
\item{df}{data.frame}

\item{cell_ids}{integer}

\item{model}{formula}

\item{n_class}{integer}

\item{genome}{character}

\item{gen}{integer}
}
\value{
list
}
\description{
Run the nth generation of models
}
