% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGeneAlgorithm.R
\name{runGeneAlgorithm}
\alias{runGeneAlgorithm}
\title{Run genetic algorithm}
\usage{
runGeneAlgorithm(
  ttv,
  cell_ids,
  organism_models,
  n_class = 4,
  n_gen = 2000,
  max_mute_intensity = 0.5,
  mute_decay = 2,
  pop = 52,
  grid,
  resolution
)
}
\arguments{
\item{ttv}{list}

\item{cell_ids}{character}

\item{organism_models}{list}

\item{n_class}{integer}

\item{n_gen}{integer}

\item{max_mute_intensity}{numeric}

\item{mute_decay}{numeric}

\item{pop}{integer}

\item{grid}{sf list}

\item{resolution}{integer}
}
\value{
list
}
\description{
Run genetic algorithm
}
