% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{seeGrid}
\alias{seeGrid}
\title{see the grid overlay for a gdf created with a given resolution}
\usage{
seeGrid(gdf, map, fill_value = "ln_sale_price")
}
\arguments{
\item{gdf}{data.frame}

\item{map}{ggmap}

\item{fill_value}{character}
}
\value{
ggplot
}
\description{
see the grid overlay for a gdf created with a given resolution
}
