% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{seeShp}
\alias{seeShp}
\title{see the shp overlay for a gdf created with a given resolution}
\usage{
seeShp(gdf, shp, map, fill_value = "ln_sale_price", discrete_color = FALSE)
}
\arguments{
\item{gdf}{data.frame}

\item{shp}{sp}

\item{map}{ggmap}

\item{fill_value}{character}

\item{discrete_color}{boolean}
}
\value{
ggplot
}
\description{
see the shp overlay for a gdf created with a given resolution
}
