% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortSpagClass.R
\name{sortSpagClass}
\alias{sortSpagClass}
\title{Attach and sort spag class from organism to object to gdf object}
\usage{
sortSpagClass(gdf, organism, mf)
}
\arguments{
\item{gdf}{data.frame}

\item{organism}{list}

\item{mf}{data.frame}
}
\value{
data.frame
}
\description{
Attach and sort spag class from organism to object to gdf object
}
