% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TTV}
\alias{TTV}
\title{Split one data.frame into list of train/test/validate data.frames}
\usage{
TTV(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
list
}
\description{
Split one data.frame into list of train/test/validate data.frames
}
