% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOrganismModels.R
\name{makeOrganismModels}
\alias{makeOrganismModels}
\title{Return models and bounds for input into runGeneAlgorithm}
\usage{
makeOrganismModels(gdf, mf, time_var = "sale_quarter")
}
\arguments{
\item{gdf}{data.frame}

\item{mf}{data.frame}

\item{time_var}{character}
}
\value{
list
}
\description{
Return models and bounds for input into runGeneAlgorithm
}
